Rename file in snake case
