+++ ditto_web_api/DittoWebApi/src/models/file_information.py
